+++ nushell/filter.py
@@ -11,7 +11,6 @@ from nushell.plugin import PluginBase
 import copy
 import fileinput
 import json
-import sys
 
 
 class FilterPlugin(PluginBase):
@@ -41,14 +40,16 @@ class FilterPlugin(PluginBase):
         else:
             return list(self.params["item"]["Primitive"].values())[0]
 
-    def print_primitive_response(self, value, primitive_type):
+
+    def print_primitive_response(self, value, primitive_type, 
+                                       return_response=False):
         '''a base function to print a good response with an updated
            value for a primitive type.
-
            Parameters
            ==========
            value: The primitive value, expected to match the type
            primitive_type: one of Int or String
+           return_response: if True, just return (don't print)
         '''
         primitive_type = self._camel_case(primitive_type)
         item = {"Primitive": {primitive_type: value}}
@@ -56,7 +57,13 @@ class FilterPlugin(PluginBase):
         # Get the original complete passed parameters to update
         response = copy.deepcopy(self.params)
         response["item"] = item
-        self.print_good_response([{"Ok": {"Value": response}}])
+        response = [{"Ok": {"Value": response}}]
+
+        # For testing, we might just want to return response
+        if return_response:
+            return response
+
+        self.print_good_response(response)
 
 
     def print_int_response(self, value):
@@ -64,6 +71,42 @@ class FilterPlugin(PluginBase):
         
     def print_string_response(self, value):
         return self.print_primitive_response(value, "String")
+
+    def test(self, runFilter, line):
+        '''given a line, test the response
+        '''
+        method = line.get("method")
+
+        # Store raw parameters for the plugin memory, only if not end filter
+        if method != "end_filter":
+            self.params = line.get('params', {})
+
+        # Case 1: Nu is asking for the config to discover the plugin
+        if method == "config":
+            plugin_config = self.get_config()
+            return self.get_good_response(plugin_config)
+
+        elif method == "begin_filter":
+
+            # Arguments only show up for begin_filter
+            self.args = self.parse_params(self.params)
+            return self.get_good_response([])
+
+        # End filter can end the filter, OR call a custom sink function
+        elif method == "end_filter":
+
+            # If the user wants help, return the help and break
+            if "help" in self.args:
+
+                # We need to update so name_tag is tag (not logical I know)
+                self.params['tag'] = self.params.get('name_tag', self.getTag())
+                return self.print_primitive_response(self.get_help(), "String", True)
+
+            return self.get_good_response([])
+
+        # Run the filter, passing the unparsed params
+        elif method == "filter":
+            return runFilter(self, self.args)
 
 
     def run(self, runFilter):